# Implementation/sc_server/swaggerAPIDoc.yaml
openapi: 3.0.1
info:
  title: API Documentation

  version: "1.0"
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /notification-api/private/send-token:
    post:
      tags:
      - notification-controller
      operationId: sendTokenNotification
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationTokenPOST'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                enum:
                - 200 OK
...
  /notification-api/private/test-notification:
    get:
      tags:
      - notification-controller
      operationId: testNotification
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                enum:
                  - 200 OK
      deprecated: true
  /notification-api/private/get-notifications:
    get:
      tags:
      - notification-controller
      operationId: getNotifications
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDTO'
...
components:
  schemas:
...
    NotificationTokenPOST:
      type: object
      properties:
        notificationToken:
          type: string
    NotificationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        body:
          type: string
...